index.tsxを削除しpage.tsxに統合
