+++ src/features/index.js
@@ -1,4 +1,3 @@
 import LoginButton from "./loginButton";
-import XShareButton from "./xShareButton"
 
-export { LoginButton, XShareButton }
+export { LoginButton }
